Globorisk risk calculator

Computes the 10-year risk of CVD using globorisk, a prediction model for the risk of cardiovascular disease in 182 countries.

globorisk(
  sex,
  age,
  sbp,
  tc = NA,
  dm = NA,
  smk,
  bmi = NA,
  iso,
  year,
  time = 10,
  version = c("lab", "office", "fatal"),
  type = "risk"
)

Arguments

sex
biological sex (0 = man, 1 = woman)

age
patient age (years)

sbp
systolic blood pressure (mmHg)

tc
total cholesterol (mmol/L)

dm
diabetes mellitus (0 = no, 1 = yes)

smk
current smoker (0 = no, 1 = yes)

bmi
body mass index (kg/m^2)

iso
ISO code for country of interest

year
baseline year

time
follow up time (default is 10-years)

version
calculator version, options are 'lab', 'office', or 'fatal'

type
output type, options are 'risk', 'survival', or 'all'

Value

If type = 'risk' output is a vector with estimated 10-year risk for CVD, if type = 'survival' output is a vector with estimated 10-year survival, if type = 'all' output is a data.frame with intermediate calculations

References

Ueda, Peter, Mark Woodward, Yuan Lu, Kaveh Hajifathalian, Rihab Al-Wotayan, Carlos A. Aguilar-Salinas, Alireza Ahmadvand, et al. "Laboratory-Based and Office-Based Risk Scores and Charts to Predict 10-Year Risk of Cardiovascular Disease in 182 Countries: A Pooled Analysis of Prospective Cohorts and Health Surveys." The Lancet Diabetes & Endocrinology 5, no. 3 (March 1, 2017): 196–213. https://doi.org/10.1016/S2213-8587(17)30015-3.

Hajifathalian, Kaveh, Peter Ueda, Yuan Lu, Mark Woodward, Alireza Ahmadvand, Carlos A Aguilar-Salinas, Fereidoun Azizi, et al. “A Novel Risk Score to Predict Cardiovascular Disease Risk in National Populations (Globorisk): A Pooled Analysis of Prospective Cohorts and Health Examination Surveys.” The Lancet Diabetes & Endocrinology 3, no. 5 (May 1, 2015): 339–55. https://doi.org/10.1016/S2213-8587(15)00081-9.

Examples

library(globorisk)

globorisk(
  sex = c(1, 0, 0),
  age = c(52, 60, 65),
  sbp = c(140, 160, 170),
  tc = c(4.5, 5, 5),
  dm = c(1, 1, 1),
  smk = c(0, 1, 1),
  iso = c("AFG", "AFG", "USA"),
  year = c(2000, 2000, 2020),
  version = "lab",
  type = "risk"
)

#> [1] 0.3323533 0.6000554 0.2284217